Type:Bm Descr:patch mail_mime - corrections E_NOTICE
git-svn-id: https://subversion.renater.fr/lodel/trunk@4744 9f73b02b-de1c-0410-bbc7-d36e22b4c400

+++ lodel/scripts/logic/class.data.php
@@ -1403,7 +1403,7 @@ class DataLogic
 	 */
 	private function _cleanDatabase() {
 		global $db;
-		if(!is_array($this->_existingTables)) return;
+		if(empty($this->_existingTables) || !is_array($this->_existingTables)) return;
 		foreach($this->_existingTables as $table=>$k) {
 			if(FALSE !== strpos($table, '__oldME')) {
 				$db->execute("DROP TABLE `{$table}`");
